Catboost ML
Today was using Pycaret for a classification task. Originally pycaret predicted that random forest would be the best model to use when classifying who may die from heart failure, however, removing age and diabetes attribute had pycaret say that Catboost would be the more accurate algorithm. May do this project again but with ALL attirubtes kept and using random forest to see if the results are similar or not.